+++ src/types/add_chat_members.rs
@@ -2,7 +2,7 @@ use crate::errors::*;
 use crate::types::*;
 use uuid::Uuid;
 
-/// Adds multiple new members to a chat. Currently this option is only available for supergroups and channels. This option can't be used to join a chat. Members can't be added to a channel if it has more than 200 members. Members will not be added until the chat state has been synchronized with the server
+/// Adds multiple new members to a chat. Currently, this method is only available for supergroups and channels. This method can't be used to join a chat. Members can't be added to a channel if it has more than 200 members
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct AddChatMembers {
     #[doc(hidden)]
@@ -12,8 +12,8 @@ pub struct AddChatMembers {
     client_id: Option<i32>,
     /// Chat identifier
     chat_id: i64,
-    /// Identifiers of the users to be added to the chat
-    user_ids: Vec<i32>,
+    /// Identifiers of the users to be added to the chat. The maximum number of added users is 20 for supergroups and 100 for channels
+    user_ids: Vec<i64>,
 
     #[serde(rename(serialize = "@type"))]
     td_type: String,
@@ -49,7 +49,7 @@ impl AddChatMembers {
         self.chat_id
     }
 
-    pub fn user_ids(&self) -> &Vec<i32> {
+    pub fn user_ids(&self) -> &Vec<i64> {
         &self.user_ids
     }
 }
@@ -69,7 +69,7 @@ impl RTDAddChatMembersBuilder {
         self
     }
 
-    pub fn user_ids(&mut self, user_ids: Vec<i32>) -> &mut Self {
+    pub fn user_ids(&mut self, user_ids: Vec<i64>) -> &mut Self {
         self.inner.user_ids = user_ids;
         self
     }
