+++ src/types/add_sticker_to_set.rs
@@ -11,12 +11,10 @@ pub struct AddStickerToSet {
     #[serde(rename(serialize = "@client_id", deserialize = "@client_id"))]
     client_id: Option<i32>,
     /// Sticker set owner
-    user_id: i32,
+    user_id: i64,
     /// Sticker set name
     name: String,
     /// Sticker to add to the set
-
-    #[serde(skip_serializing_if = "InputSticker::_is_default")]
     sticker: InputSticker,
 
     #[serde(rename(serialize = "@type"))]
@@ -49,7 +47,7 @@ impl AddStickerToSet {
         RTDAddStickerToSetBuilder { inner }
     }
 
-    pub fn user_id(&self) -> i32 {
+    pub fn user_id(&self) -> i64 {
         self.user_id
     }
 
@@ -72,7 +70,7 @@ impl RTDAddStickerToSetBuilder {
         self.inner.clone()
     }
 
-    pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+    pub fn user_id(&mut self, user_id: i64) -> &mut Self {
         self.inner.user_id = user_id;
         self
     }
