+++ src/types/chat_statistics_message_sender_info.rs
@@ -11,10 +11,10 @@ pub struct ChatStatisticsMessageSenderInfo {
     #[serde(rename(serialize = "@client_id", deserialize = "@client_id"))]
     client_id: Option<i32>,
     /// User identifier
-    user_id: i32,
+    user_id: i64,
     /// Number of sent messages
     sent_message_count: i32,
-    /// Average number of characters in sent messages
+    /// Average number of characters in sent messages; 0 if unknown
     average_character_count: i32,
 }
 
@@ -40,7 +40,7 @@ impl ChatStatisticsMessageSenderInfo {
         RTDChatStatisticsMessageSenderInfoBuilder { inner }
     }
 
-    pub fn user_id(&self) -> i32 {
+    pub fn user_id(&self) -> i64 {
         self.user_id
     }
 
@@ -63,7 +63,7 @@ impl RTDChatStatisticsMessageSenderInfoBuilder {
         self.inner.clone()
     }
 
-    pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+    pub fn user_id(&mut self, user_id: i64) -> &mut Self {
         self.inner.user_id = user_id;
         self
     }
