+++ src/types/create_call.rs
@@ -11,10 +11,10 @@ pub struct CreateCall {
     #[serde(rename(serialize = "@client_id", deserialize = "@client_id"))]
     client_id: Option<i32>,
     /// Identifier of the user to be called
-    user_id: i32,
-    /// Description of the call protocols supported by the application
+    user_id: i64,
+    /// The call protocols supported by the application
     protocol: CallProtocol,
-    /// True, if a video call needs to be created
+    /// Pass true to create a video call
     is_video: bool,
 
     #[serde(rename(serialize = "@type"))]
@@ -47,7 +47,7 @@ impl CreateCall {
         RTDCreateCallBuilder { inner }
     }
 
-    pub fn user_id(&self) -> i32 {
+    pub fn user_id(&self) -> i64 {
         self.user_id
     }
 
@@ -70,7 +70,7 @@ impl RTDCreateCallBuilder {
         self.inner.clone()
     }
 
-    pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+    pub fn user_id(&mut self, user_id: i64) -> &mut Self {
         self.inner.user_id = user_id;
         self
     }
