+++ src/types/game_high_score.rs
@@ -13,7 +13,7 @@ pub struct GameHighScore {
     /// Position in the high score table
     position: i32,
     /// User identifier
-    user_id: i32,
+    user_id: i64,
     /// User score
     score: i32,
 }
@@ -44,7 +44,7 @@ impl GameHighScore {
         self.position
     }
 
-    pub fn user_id(&self) -> i32 {
+    pub fn user_id(&self) -> i64 {
         self.user_id
     }
 
@@ -68,7 +68,7 @@ impl RTDGameHighScoreBuilder {
         self
     }
 
-    pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+    pub fn user_id(&mut self, user_id: i64) -> &mut Self {
         self.inner.user_id = user_id;
         self
     }
