+++ src/types/get_basic_group.rs
@@ -11,7 +11,7 @@ pub struct GetBasicGroup {
     #[serde(rename(serialize = "@client_id", deserialize = "@client_id"))]
     client_id: Option<i32>,
     /// Basic group identifier
-    basic_group_id: i32,
+    basic_group_id: i64,
 
     #[serde(rename(serialize = "@type"))]
     td_type: String,
@@ -43,7 +43,7 @@ impl GetBasicGroup {
         RTDGetBasicGroupBuilder { inner }
     }
 
-    pub fn basic_group_id(&self) -> i32 {
+    pub fn basic_group_id(&self) -> i64 {
         self.basic_group_id
     }
 }
@@ -58,7 +58,7 @@ impl RTDGetBasicGroupBuilder {
         self.inner.clone()
     }
 
-    pub fn basic_group_id(&mut self, basic_group_id: i32) -> &mut Self {
+    pub fn basic_group_id(&mut self, basic_group_id: i64) -> &mut Self {
         self.inner.basic_group_id = basic_group_id;
         self
     }
