+++ src/types/get_chat_event_log.rs
@@ -20,10 +20,10 @@ pub struct GetChatEventLog {
     from_event_id: i64,
     /// The maximum number of events to return; up to 100
     limit: i32,
-    /// The types of events to return. By default, all types will be returned
+    /// The types of events to return; pass null to get chat events of all types
     filters: ChatEventLogFilters,
     /// User identifiers by which to filter events. By default, events relating to all users will be returned
-    user_ids: Vec<i32>,
+    user_ids: Vec<i64>,
 
     #[serde(rename(serialize = "@type"))]
     td_type: String,
@@ -75,7 +75,7 @@ impl GetChatEventLog {
         &self.filters
     }
 
-    pub fn user_ids(&self) -> &Vec<i32> {
+    pub fn user_ids(&self) -> &Vec<i64> {
         &self.user_ids
     }
 }
@@ -115,7 +115,7 @@ impl RTDGetChatEventLogBuilder {
         self
     }
 
-    pub fn user_ids(&mut self, user_ids: Vec<i32>) -> &mut Self {
+    pub fn user_ids(&mut self, user_ids: Vec<i64>) -> &mut Self {
         self.inner.user_ids = user_ids;
         self
     }
