+++ src/types/get_groups_in_common.rs
@@ -11,7 +11,7 @@ pub struct GetGroupsInCommon {
     #[serde(rename(serialize = "@client_id", deserialize = "@client_id"))]
     client_id: Option<i32>,
     /// User identifier
-    user_id: i32,
+    user_id: i64,
     /// Chat identifier starting from which to return chats; use 0 for the first request
     offset_chat_id: i64,
     /// The maximum number of chats to be returned; up to 100
@@ -47,7 +47,7 @@ impl GetGroupsInCommon {
         RTDGetGroupsInCommonBuilder { inner }
     }
 
-    pub fn user_id(&self) -> i32 {
+    pub fn user_id(&self) -> i64 {
         self.user_id
     }
 
@@ -70,7 +70,7 @@ impl RTDGetGroupsInCommonBuilder {
         self.inner.clone()
     }
 
-    pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+    pub fn user_id(&mut self, user_id: i64) -> &mut Self {
         self.inner.user_id = user_id;
         self
     }
